Update addendum parsing
